pyproject.toml
[project]
name = "compare-runners"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-gitlab>=6.0.0",
    "tabulate>=0.9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/compare_runners"]

[project.scripts]
main = "compare_runners.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
